add granular controls for web push notification setting types, toggle visibility depending on subscription in browser, unsubscribe on disable browser notifications via app
