+++ shared/graphql/fragments/user/userSettings.js
@@ -7,21 +7,27 @@ export type UserSettingsType = {
       types: {
         newMessageInThreads: {
           email: boolean,
+          web: boolean,
         },
         newDirectMessage: {
           email: boolean,
+          web: boolean,
         },
         newThreadCreated: {
           email: boolean,
+          web: boolean,
         },
         weeklyDigest: {
           email: boolean,
+          web: boolean,
         },
         dailyDigest: {
           email: boolean,
+          web: boolean,
         },
         newMention: {
           email: boolean,
+          web: boolean,
         },
       },
     },
@@ -35,21 +41,27 @@ export default gql`
         types {
           newMessageInThreads {
             email
+            web
           }
           newDirectMessage {
             email
+            web
           }
           newThreadCreated {
             email
+            web
           }
           weeklyDigest {
             email
+            web
           }
           dailyDigest {
             email
+            web
           }
           newMention {
             email
+            web
           }
         }
       }
